[docker] One more attempt to make per-container reapers
We had an attempt to make the reapers per-container, and that didn't
quite work.

This change now makes sure that every process creates a listener in a
random-ish path and then inherits that to its children. No instance
should ever stomp over the socket of another. And if a chain of
processes set their own subprocess listeners, those are concatenated
(although it's not strictly necessary, because setting a subreaper lower
down the process tree should catch processes before their parents get a
go at it).